extract TwigRenderer
