+++ src/ApiDoc.php
@@ -10,14 +10,12 @@ use BEAR\Resource\RenderInterface;
 use BEAR\Resource\ResourceInterface;
 use BEAR\Resource\ResourceObject;
 use BEAR\Resource\TransferInterface;
-use LogicException;
-use manuelodelain\Twig\Extension\LinkifyExtension;
-use Ray\Di\Di\Inject;
-use Ray\Di\Di\Named;
-use Twig_Extension_Debug;
 use function file_get_contents;
 use function json_decode;
 use function json_encode;
+use LogicException;
+use Ray\Di\Di\Inject;
+use Ray\Di\Di\Named;
 use function str_replace;
 
 class ApiDoc extends ResourceObject
@@ -84,26 +82,7 @@ class ApiDoc extends ResourceObject
     public function setRenderer(RenderInterface $renderer)
     {
         unset($renderer);
-        $this->renderer = new class($this->template) implements RenderInterface {
-            private $template;
-
-            public function __construct(array $template)
-            {
-                $this->template = $template;
-            }
-
-            public function render(ResourceObject $ro)
-            {
-                $ro->headers['content-type'] = 'text/html; charset=utf-8';
-                $twig = new \Twig_Environment(new \Twig_Loader_Array($this->template), ['debug' => true]);
-                $twig->addExtension(new Twig_Extension_Debug);
-                $twig->addExtension(new RefLinkExtention);
-                $twig->addExtension(new LinkifyExtension);
-                $ro->view = $twig->render('index', $ro->body);
-
-                return $ro->view;
-            }
-        };
+        $this->renderer = new TwigRenderer($this->template);
 
         return $this;
     }
